feat(dot): add command
